refactor: Updated SystemResponse.ts,errorHandler.ts,IUserModel.ts,UserModel.ts,UserRepository.ts and seedData.ts
