+++ src/controllers/user/UserController.ts
@@ -1,8 +1,6 @@
 import { Request, Response } from 'express';
-import UserRepository from '../../repositories/users/UserRepository';
+import UserRepository from '../../repositories/user/UserRepository';
 import SystemResponse from '../../libs/SystemResponse';
-import * as mongoose from 'mongoose';
-
 
 class UserController {
 
@@ -22,8 +20,8 @@ class UserController {
 
         this.userRepository.get().then(data => {
             SystemResponse.success(res, data, 'Trainee Data Founded');
-        }).catch(err => {
-            throw err;
+        }).catch(error => {
+            SystemResponse.failure(res, error);
         });
 
     }
@@ -32,16 +30,16 @@ class UserController {
         const { id, dataToUpdate } = req.body;
         this.userRepository.update(id, dataToUpdate).then(data => {
             SystemResponse.success(res, data, 'Trainee Data Updated');
-        }).catch(err => {
-            throw err;
+        }).catch(error => {
+            SystemResponse.failure(res, error);
         });
     }
 
     post = (req: Request, res: Response): void => {
         this.userRepository.create(req.body).then(user => {
-             SystemResponse.success(res, user, 'Trainee Data Added');
+            SystemResponse.success(res, user, 'Trainee Data Added');
         }).catch(error => {
-            throw error;
+            SystemResponse.failure(res, error);
         });
 
     }
@@ -51,7 +49,7 @@ class UserController {
         this.userRepository.delete(id).then(user => {
             SystemResponse.success(res, user, 'Trainee Data Deleted');
         }).catch(error => {
-            throw error;
+            SystemResponse.failure(res, error);
         });
     }
