+++ src/libs/Database.ts
@@ -10,7 +10,6 @@ class DataBase {
                     reject(err);
                 }
                 resolve();
-                console.log('DB Connected Successfully');
                 seedData();
             }).catch(error => (console.log(error)));
             console.log('Data Base Connected Successfully');
