+++ src/repositories/user/UserModel.ts
@@ -1,10 +1,10 @@
 import * as mongoose from 'mongoose';
-import { IUserModel } from './IUsermodel';
+import { IUserModel } from './IUserModel';
 import UserSchema from './UserSchema';
 
 export const userSchema = new UserSchema({
     collection: 'user'
 });
 
-export const userModel: mongoose.model<IUserModel> = mongoose.model<IUserModel>(
+export const userModel: mongoose.Model<IUserModel> = mongoose.model<IUserModel>(
     'user', userSchema, 'Users', true);
